refactor(Core UI): :recycle: Skloader as core component
