+++ __tests__/pages/ErrorConsole.spec.tsx
@@ -4,9 +4,9 @@ import { render, screen } from '@testing-library/react';
 import { Server } from 'miragejs';
 
 import { loadMockServer } from '../../mocks/server';
+import LoadingPage from '../../src/core/components/SkLoading';
 import { Wrapper } from '../../src/core/components/Wrapper';
 import ErrorConsole from '../../src/pages/shared/Errors/Console';
-import LoadingPage from '../../src/pages/shared/Loading';
 
 describe('Begin testing the Error Console view', () => {
   let server: Server;
