+++ __tests__/pages/Metrics/TcpConnection.spec.tsx
@@ -8,8 +8,8 @@ import { loadMockServer } from '../../../mocks/server';
 import * as PrometheusAPIModule from '../../../src/API/Prometheus.api';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/config';
 import { getTestsIds } from '../../../src/config/testIds';
+import LoadingPage from '../../../src/core/components/SkLoading';
 import { Wrapper } from '../../../src/core/components/Wrapper';
-import LoadingPage from '../../../src/pages/shared/Loading';
 import TcpConnection from '../../../src/pages/shared/Metrics/components/TcpConnection';
 import { MetricsLabels } from '../../../src/pages/shared/Metrics/Metrics.enum';
 import { ProcessResponse } from '../../../src/types/REST.interfaces';
