+++ __tests__/pages/Metrics/Traffic.spec.tsx
@@ -7,8 +7,8 @@ import processesData from '../../../mocks/data/PROCESSES.json';
 import { loadMockServer } from '../../../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/config';
 import { getTestsIds } from '../../../src/config/testIds';
+import LoadingPage from '../../../src/core/components/SkLoading';
 import { Wrapper } from '../../../src/core/components/Wrapper';
-import LoadingPage from '../../../src/pages/shared/Loading';
 import Traffic from '../../../src/pages/shared/Metrics/components/Traffic';
 import { MetricsLabels } from '../../../src/pages/shared/Metrics/Metrics.enum';
 import * as MetricsModule from '../../../src/pages/shared/Metrics/services/index';
