+++ __tests__/pages/Processes/PairsList.spec.tsx
@@ -3,17 +3,17 @@ import { Suspense } from 'react';
 import { render, screen, waitForElementToBeRemoved } from '@testing-library/react';
 import { Server } from 'miragejs';
 
-import { SkEmptyDataLabels } from '../../../src/core/components/SkEmptyData';
 
 import processesPairsData from '../../../mocks/data/PROCESS_PAIRS.json';
 import processesData from '../../../mocks/data/PROCESSES.json';
 import { loadMockServer } from '../../../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/config';
 import { getTestsIds } from '../../../src/config/testIds';
+import { SkEmptyDataLabels } from '../../../src/core/components/SkEmptyData';
+import LoadingPage from '../../../src/core/components/SkLoading';
 import { Wrapper } from '../../../src/core/components/Wrapper';
 import PairsList from '../../../src/pages/Processes/components/PairsList';
 import { ProcessesLabels, ProcessesRoutesPaths } from '../../../src/pages/Processes/Processes.enum';
-import LoadingPage from '../../../src/pages/shared/Loading';
 import { ProcessResponse, PairsResponse } from '../../../src/types/REST.interfaces';
 
 const data = processesData.results[7] as ProcessResponse;
