+++ __tests__/pages/Processes/Process.spec.tsx
@@ -8,10 +8,10 @@ import processesData from '../../../mocks/data/PROCESSES.json';
 import { loadMockServer } from '../../../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/config';
 import { getTestsIds } from '../../../src/config/testIds';
+import LoadingPage from '../../../src/core/components/SkLoading';
 import { Wrapper } from '../../../src/core/components/Wrapper';
 import { ProcessesLabels } from '../../../src/pages/Processes/Processes.enum';
 import Process from '../../../src/pages/Processes/views/Process';
-import LoadingPage from '../../../src/pages/shared/Loading';
 import { ProcessResponse } from '../../../src/types/REST.interfaces';
 
 const processResult = processesData.results[0] as ProcessResponse;
