+++ __tests__/pages/Services/PairList.spec.tsx
@@ -3,15 +3,15 @@ import { Suspense } from 'react';
 import { render, screen, waitForElementToBeRemoved } from '@testing-library/react';
 import { Server } from 'miragejs';
 
-import ProcessServerList from '../../../src/pages/Services/components/ProcessServerList';
 
 import biFlowData from '../../../mocks/data/FLOW_PAIRS.json';
 import servicesData from '../../../mocks/data/SERVICES.json';
 import { loadMockServer } from '../../../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/config';
 import { getTestsIds } from '../../../src/config/testIds';
+import LoadingPage from '../../../src/core/components/SkLoading';
 import { Wrapper } from '../../../src/core/components/Wrapper';
-import LoadingPage from '../../../src/pages/shared/Loading';
+import ProcessServerList from '../../../src/pages/Services/components/ProcessServerList';
 
 const servicesResults = servicesData.results;
 const biFlowResults = biFlowData.results;
