+++ __tests__/pages/Services/Services.spec.tsx
@@ -7,9 +7,9 @@ import servicesData from '../../../mocks/data/SERVICES.json';
 import { loadMockServer } from '../../../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/config';
 import { getTestsIds } from '../../../src/config/testIds';
+import LoadingPage from '../../../src/core/components/SkLoading';
 import { Wrapper } from '../../../src/core/components/Wrapper';
 import Services from '../../../src/pages/Services/views/Services';
-import LoadingPage from '../../../src/pages/shared/Loading';
 
 const servicesResults = servicesData.results;
 const service = servicesResults[0];
