+++ __tests__/pages/Services/TcpConnections.spec.tsx
@@ -3,15 +3,15 @@ import { Suspense } from 'react';
 import { render, screen, waitForElementToBeRemoved } from '@testing-library/react';
 import { Server } from 'miragejs';
 
-import TcpConnections from '../../../src/pages/Services/components/TcpConnections';
 
 import biFlowData from '../../../mocks/data/FLOW_PAIRS.json';
 import servicesData from '../../../mocks/data/SERVICES.json';
 import { loadMockServer } from '../../../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/config';
 import { getTestsIds } from '../../../src/config/testIds';
+import LoadingPage from '../../../src/core/components/SkLoading';
 import { Wrapper } from '../../../src/core/components/Wrapper';
-import LoadingPage from '../../../src/pages/shared/Loading';
+import TcpConnections from '../../../src/pages/Services/components/TcpConnections';
 
 const servicesResults = servicesData.results;
 const tcpBiFlowOpen = biFlowData.results[5];
