+++ src/core/components/SkLoading/Loading.enum.ts
